changed the dialogue location and remove unnecessary stuff
